refactor: rename src/ui to src/tui
