+++ src/tui/components/HelpRow.tsx
