+++ src/tui/components/LoadingRow.tsx
