+++ src/tui/components/PathSetupForm.tsx
