+++ src/tui/components/ProgressBar.tsx
